Implement test editor
