+++ AuroraEditor/Base/Documents/UI/WorkspaceCodeFileView.swift
@@ -7,6 +7,7 @@
 
 import SwiftUI
 import UniformTypeIdentifiers
+import CodeEditorView
 
 struct WorkspaceCodeFileView: View {
     var windowController: NSWindowController
@@ -58,7 +59,7 @@ struct WorkspaceCodeFileView: View {
                                 }
                             })
                     } else {
-                        aeCodeView(fileItem, for: item)
+                        codeEditorView(fileItem, for: item)
                             .splitView(availablePositions: [.top, .bottom, .center, .leading, .trailing],
                                        proposalPosition: $dropProposal,
                                        margin: 15,
@@ -117,6 +118,21 @@ struct WorkspaceCodeFileView: View {
     }
 
     @ViewBuilder
+    private func codeEditorView(
+        _ codeFile: CodeFileDocument,
+        for item: FileSystemClient.FileItem
+    ) -> some View {
+        // TODO: Wesley - implement new editor.
+        CodeEditorViewWrapper(codeFile: codeFile)
+            .safeAreaInset(edge: .top, spacing: 0) {
+                VStack(spacing: 0) {
+                    BreadcrumbsView(file: item, tappedOpenFile: workspace.openTab(item:))
+                    Divider()
+                }
+            }
+    }
+
+    @ViewBuilder
     private func imageFileView(
         _ otherFile: CodeFileDocument,
         for item: FileSystemClient.FileItem
@@ -148,71 +164,5 @@ struct WorkspaceCodeFileView: View {
     var body: some View {
         codeView
             .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-}
-
-// MARK: AECodeView
-public struct AECodeView: View {
-    @ObservedObject
-    private var codeFile: CodeFileDocument
-
-    @ObservedObject
-    private var prefs: AppPreferencesModel = .shared
-
-    @Environment(\.colorScheme)
-    private var colorScheme
-
-    private let editable: Bool
-
-    public init(codeFile: CodeFileDocument, editable: Bool = true) {
-        self.codeFile = codeFile
-        self.editable = editable
-    }
-
-    @State
-    private var selectedTheme = ThemeModel.shared.selectedTheme ?? ThemeModel.shared.themes.first!
-
-    @State
-    private var font: NSFont = {
-        let size = AppPreferencesModel.shared.preferences.textEditing.font.size
-        let name = AppPreferencesModel.shared.preferences.textEditing.font.name
-        return NSFont(name: name, size: Double(size)) ?? NSFont.monospacedSystemFont(ofSize: 12, weight: .regular)
-    }()
-
-    private var themeString: String? {
-        return ThemeModel.shared.selectedTheme?.highlightrThemeString
-    }
-
-    private func getLanguage() -> CodeLanguage? {
-        if let url = codeFile.fileURL {
-            return .detectLanguageFromUrl(url: url)
-        } else {
-            return .default
-        }
-    }
-
-    public var body: some View {
-        AuroraEditorTextView(
-            $codeFile.content,
-            font: $font,
-            tabWidth: $prefs.preferences.textEditing.defaultTabWidth,
-            lineHeight: .constant(1.2),
-            language: getLanguage(),
-            themeString: themeString
-        )
-        .id(codeFile.fileURL)
-        .background(selectedTheme.editor.background.swiftColor)
-        .disabled(!editable)
-        .frame(maxHeight: .infinity)
-        .onChange(of: ThemeModel.shared.selectedTheme) { newValue in
-            guard let theme = newValue else { return }
-            self.selectedTheme = theme
-        }
-        .onChange(of: prefs.preferences.textEditing.font) { _ in
-            font = NSFont(
-                name: prefs.preferences.textEditing.font.name,
-                size: Double(prefs.preferences.textEditing.font.size)
-            ) ?? .monospacedSystemFont(ofSize: 12, weight: .regular)
-        }
     }
 }
