+++ Package.resolved
@@ -11,6 +11,15 @@
         }
       },
       {
+        "package": "CodeEditorView",
+        "repositoryURL": "https://github.com/mchakravarty/CodeEditorView",
+        "state": {
+          "branch": null,
+          "revision": "325759c1cc2873fd5eba0f8d4eedd0233ac9eca7",
+          "version": "0.10.0"
+        }
+      },
+      {
         "package": "SwiftTerm",
         "repositoryURL": "https://github.com/migueldeicaza/SwiftTerm.git",
         "state": {
